+++ src/components/map/map-legend.js
@@ -45,11 +45,11 @@ const StyledMapControlLegend = styled.div`
     font-size: 11px;
     padding-right: ${props => props.theme.mapControl.padding}px;
   }
-  
-  .legend--layer__title { 
+
+  .legend--layer__title {
     padding-right: ${props => props.theme.mapControl.padding}px;
   }
-  
+
   .legend--layer_by {
     color: ${props => props.theme.subtextColor};
   }
@@ -113,13 +113,13 @@ const MultiColorLegend = ({layer, width}) => {
 const MapLegend = ({layers}) => (
   <div>
     {layers.map((layer, index) => {
-      const colorChannelConfig = layer.getVisualChannelDescription('color');
-      const enableColorBy = colorChannelConfig.measure;
-      const width = DIMENSIONS.mapControl.width - 2 * DIMENSIONS.mapControl.padding;
-
       if (!layer.isValidToSave()) {
         return null;
       }
+
+      const colorChannelConfig = layer.getVisualChannelDescription('color');
+      const enableColorBy = colorChannelConfig.measure;
+      const width = DIMENSIONS.mapControl.width - 2 * DIMENSIONS.mapControl.padding;
 
       return (
         <StyledMapControlLegend