+++ src/layers/point-layer/point-layer.js
@@ -22,6 +22,7 @@ import Layer from '../base-layer';
 import memoize from 'lodash.memoize';
 import {TextLayer} from 'deck.gl';
 import ScatterplotBrushingLayer from 'deckgl-layers/scatterplot-brushing-layer/scatterplot-brushing-layer';
+import uniq from 'lodash.uniq';
 import {hexToRgb} from 'utils/color-utils';
 import PointLayerIcon from './point-layer-icon';
 import {DEFAULT_LAYER_COLOR} from 'constants/default-settings';
@@ -34,6 +35,10 @@ export const pointPosAccessor = ({lat, lng, altitude}) => d => [
 
 export const pointPosResolver = ({lat, lng, altitude}) =>
   `${lat.fieldIdx}-${lng.fieldIdx}-${altitude ? altitude.fieldIdx : 'z'}`;
+
+export const pointLabelAccessor = textLabel => d => String(d.data[textLabel.field.tableFieldIndex - 1]);
+export const pointLabelResolver = textLabel => textLabel.field && textLabel.field.tableFieldIndex;
+
 export const pointRequiredColumns = ['lat', 'lng'];
 export const pointOptionalColumns = ['altitude'];
 
@@ -54,6 +59,7 @@ export default class PointLayer extends Layer {
 
     this.registerVisConfig(pointVisConfigs);
     this.getPosition = memoize(pointPosAccessor, pointPosResolver);
+    this.getText = memoize(pointLabelAccessor, pointLabelResolver);
   }
 
   get type() {
@@ -142,6 +148,7 @@ export default class PointLayer extends Layer {
       sizeField,
       sizeScale,
       sizeDomain,
+      textLabel,
       visConfig: {radiusRange, fixedRadius, colorRange}
     } = this.config;
 
@@ -191,6 +198,21 @@ export default class PointLayer extends Layer {
       }, []);
     }
 
+    // get all distinct characters in the text labels
+    const getText = this.getText(textLabel);
+    let labelCharacterSet;
+    if (
+      oldLayerData &&
+      oldLayerData.labelCharacterSet &&
+      opt.sameData &&
+      oldLayerData.getText === getText
+    ) {
+      labelCharacterSet = oldLayerData.labelCharacterSet
+    } else {
+      const textLabels = textLabel.field ? data.map(getText) : [];
+      labelCharacterSet = uniq(textLabels.join(''));
+    }
+
     const getRadius = rScale ? d =>
       this.getEncodedChannelValue(rScale, d.data, sizeField) : 1;
 
@@ -199,9 +221,11 @@ export default class PointLayer extends Layer {
 
     return {
       data,
+      labelCharacterSet,
       getPosition,
       getColor,
-      getRadius
+      getRadius,
+      getText
     };
   }
 
@@ -246,9 +270,9 @@ export default class PointLayer extends Layer {
         id: this.id,
         opacity: this.config.visConfig.opacity,
         pickable: true,
-        // parameters
         parameters: {
-          depthTest: mapState.dragRotate
+          // circles will be flat on the map when the altitude column is not used
+          depthTest: this.config.columns.altitude.fieldIdx > -1
         },
 
         updateTriggers: {
@@ -276,8 +300,14 @@ export default class PointLayer extends Layer {
               getPixelOffset: this.config.textLabel.offset,
               getSize: this.config.textLabel.size,
               getTextAnchor: this.config.textLabel.anchor,
-              getText: d => String(d.data[this.config.textLabel.field.tableFieldIndex - 1]),
+              getText: data.getText,
               getColor: d => this.config.textLabel.color,
+              fp64: this.config.visConfig['hi-precision'],
+              parameters: {
+                // text will always show on top of all layers
+                depthTest: false
+              },
+              characterSet: data.labelCharacterSet,
               updateTriggers: {
                 getPosition: data.getPosition,
                 getPixelOffset: this.config.textLabel.offset,